No autoboxing at Bamboo ?!?
git-svn-id: http://svn.codehaus.org/xstream/trunk@1469 9830eeb5-ddf4-0310-9ef7-f4b9a3e3227e

+++ xstream/src/test/com/thoughtworks/acceptance/LocalConverterTest.java
@@ -86,7 +86,7 @@ public class LocalConverterTest extends AbstractAcceptanceTest {
         }
 
         public String toString(Object obj) {
-            return Integer.toHexString((Integer)obj);
+            return Integer.toHexString(((Integer)obj).intValue());
         }
     }
